Uses delegated forecaster instead for improved maintainability
